In BackendMongo.update, I overwrote the value of value with a random string so it did the update with the random string instead of the original value.
Then, when test_kvpair2.py called BackendMongo.update, the value that it read back was the value of the random string, NOT the value it said to update with.

There is a whole bunch of extraneous code in there to help me find the problem, some of which should be left in there.